Allow user to enable/disable enemy sounds.
Adds a new .txt to the settings folder called 'od-enable-enemy-sounds.txt'.

If the content inside is '1', game will behave as usual.
If the content inside is '0', game will not play sound effects coming from enemies.

More specifically, this will make that the SFX coming from enemies will have an early return in method422 in UE class, i.e. it's not that the SFX are muted, it's that they are not played, so disabling this option should improve performance.